More structure, but still a lot missing
inter_deltat update has a lot more of the structure I'm going for.
The main bit missing there is the actual pushing of values to other
objects, which I'm not quite sure how to do.

the participating objects have been implemented in a vector, with
a structure SEC_CNTRL_PARTICIPANT defined.
A skeleton for parsing an input is started but I'm not actually
sure how to work with string (char?) parsing.

the main pre deltat mode initialization is done (I believe).
Basically just zeroing derivatives.