Implemented pdispatch for diesel_dg
variable pdispatch and pdispatch_offset create a unified interface
for setting the active power output of a generator irrespective of
the control mode (as long as the control mode is supported, for
example, GGOV1 with Rselect=0 or GAST are not supported).

pdispatch is the *desired* output in p.u. on the machine basis.
pdispatch_offset is an *offset* to the desired output.

pdispatch and pdispatch_offset are *lower priority* to the direct
governor control variables. That means that if Pref or wref are set
(via a player) these values will override and update pdispatch (and
pdispatch_offset will be set to 0).

Example:
Machine base is 1 MVA

pdispatch = 0.5
pdispatch_offset = 0
The output will (in steady state) go to 500kW.

Now, change pdispatch_offset to 0.25. State is:
pdispatch = 0.5
pdispatch_offset = 0.25
The output will (in steady state) go to 750kW.

Now, change pdispatch to 0.75. State is:
pdispatch = 0.75
pdispatch_offset = 0.25 (NOTE THAT pdispatch_offset is not zeroed!!!)
The output will (in stead state) go to 1000kW.