Modification to sample handling
Sampling was previously performed as a last step, now it is done
directly after the PID output is calculated.
That is the deltas to the individual generators are not calculated
unless we are actually exporting on this step.

This is mainly done because it make some of the anti-windup conceptually
cleaner